Auth y Admin

+++ src/services/movieService.js
@@ -1,5 +1,10 @@
 const API_URL = `${import.meta.env.VITE_API_URL}/movies`;
 
+const getToken = () => {
+  const token = localStorage.getItem("token");
+  return token ? `Bearer ${token}` : null;
+};
+
 const handleResponse = async (response) => {
   const data = await response.json();
 
@@ -23,9 +28,15 @@ export const getMovieById = async (movieId) => {
 };
 
 export const createMovie = async (movieData) => {
+  const token = getToken();
+
+  if (!token) {
+    throw new Error("No autorizado");
+  }
+
   const response = await fetch(API_URL, {
     method: "POST",
-    headers: { "Content-Type": "application/json" },
+    headers: { "Content-Type": "application/json", Authorization: token },
     body: JSON.stringify(movieData),
   });
 
@@ -33,9 +44,15 @@ export const createMovie = async (movieData) => {
 };
 
 export const updateMovie = async (movieId, movieData) => {
+  const token = getToken();
+
+  if (!token) {
+    throw new Error("No autorizado");
+  }
+
   const response = await fetch(`${API_URL}/${movieId}`, {
     method: "PUT",
-    headers: { "Content-Type": "application/json" },
+    headers: { "Content-Type": "application/json", Authorization: token },
     body: JSON.stringify(movieData),
   });
 
@@ -43,8 +60,15 @@ export const updateMovie = async (movieId, movieData) => {
 };
 
 export const deleteMovie = async (movieId) => {
+  const token = getToken();
+
+  if (!token) {
+    throw new Error("No autorizado");
+  }
+
   const response = await fetch(`${API_URL}/${movieId}`, {
     method: "DELETE",
+    headers: { Authorization: token },
   });
 
   return handleResponse(response);
